Extracted the responsibility of Formatting text for display into a separate class named Formatter
